+++ packages/create/src/index.ts
@@ -1,7 +1,7 @@
+import type {CreatePackageOptions} from './types.js'
 import fs from 'node:fs/promises'
 import path from 'node:path'
 import {run} from '@sxzz/create'
-import type {CreatePackageOptions} from './types.js'
 
 /**
  * Creates a new package based on a specified template.