+++ packages/eslint-config/scripts/generate-types.ts
@@ -1,6 +1,6 @@
 import fs from 'node:fs/promises'
-import {builtinRules} from 'eslint/use-at-your-own-risk'
 import {flatConfigsToRulesDTS} from 'eslint-typegen/core'
+import {builtinRules} from 'eslint/use-at-your-own-risk'
 import {defineConfig} from '../src/define-config'
 
 const configs = await defineConfig({
@@ -27,7 +27,7 @@ const configDts = `import type {Linter} from 'eslint'
 import type {FlatConfigComposer, ResolvableFlatConfig} from 'eslint-flat-config-utils'
 import type {${rulesTypeName}} from './rules'
 
-export type {FlatConfigComposer, ResolvableFlatConfig, ${rulesTypeName}}
+export type {FlatConfigComposer, ResolvableFlatConfig}
 
 /**
  * Represents a value that resolves to one or more ESLint flat configurations.