+++ packages/eslint-config/src/plugins.ts
@@ -5,7 +5,7 @@ export async function interopDefault<T>(
   m: Awaitable<T>,
 ): Promise<T extends {default: infer U} ? U : T> {
   const resolved = await m
-  return (resolved as any).default || resolved
+  return 'default' in resolved ? interopDefault(resolved.default) : resolved
 }
 
 // @ts-expect-error - No types