+++ packages/prettier-config/tsup.config.ts
@@ -7,22 +7,20 @@ export default defineConfig([
     dts: true,
     entry: ['./src/*.ts'],
     format: ['esm'],
-    minify: true,
     outDir: 'lib',
     sourcemap: true,
     // Splitting can place code using `import.meta` into a separate chunk, which breaks the preset export mappings.
     splitting: false,
-    target: 'es2022',
   },
   {
-    entry: ['./src/prettier.config.ts'],
+    entry: {'prettier.config': './src/index.ts'},
     format: ['esm'],
     minify: true,
     // Bundle the config and dependencies into a single file.
     noExternal: [/^@bfra.me\//],
     outDir: '.',
+    shims: true,
     sourcemap: true,
-    target: 'es2022',
-    treeshake: 'smallest',
+    splitting: false,
   },
 ])