+++ packages/api-core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @bfra.me/api-core
 
+## 0.1.5
+### Patch Changes
+
+
+- Fix linter errors. ([#691](https://github.com/bfra-me/works/pull/691))
+
+## 0.1.4
+### Patch Changes
+
+
+- Fix linter errors. ([#690](https://github.com/bfra-me/works/pull/690))
+
+
+- Updated dependency `vitest` to `2.1.6`. ([#686](https://github.com/bfra-me/works/pull/686))
+  Updated dependency `@vitest/coverage-v8` to `2.1.6`.
+
 ## 0.1.3
 ### Patch Changes