+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @bfra.me/tsconfig
 
+## 0.9.3
+### Patch Changes
+
+
+- Updated dependency `type-fest` to `4.28.1`. ([#682](https://github.com/bfra-me/works/pull/682))
+
+## 0.9.2
+### Patch Changes
+
+
+- Updated dependency `type-fest` to `4.28.0`. ([#656](https://github.com/bfra-me/works/pull/656))
+
+
+- Updated dependency `type-fest` to `4.27.1`. ([#654](https://github.com/bfra-me/works/pull/654))
+
 ## 0.9.1
 ### Patch Changes